app/requirements.txt
django==3.1.6
pillow==8.1.0
django-cleanup==5.1.0
django-crispy-forms==1.11.1
django-debug-toolbar==3.2
factory-boy==3.2.0
psycopg2-binary==2.8.6
gunicorn==20.0.4
